bugs identificados, modo debug incrementado

+++ include/Sprite.h
@@ -29,7 +29,7 @@ class Sprite
         void SetPosition(int x, int y);
         void Resize(int w, int h);
         void Clip(int w, int h, int x = 0, int y = 0);
-        void SetRotationAngle(double angle);
+        void SetRotationAngle(int angle);
         void SetRotationPoint(SDL_Point pt);
         void Transform(int w, int h);
         bool IsMouseInside();
@@ -47,7 +47,7 @@ class Sprite
         SDL_Rect srcRect;
         string path;
         SDL_Point pt;
-        double angle;
+        int angle;
         Uint32 alpha;
 };
